+++ src/backend/core/urls.py
@@ -26,6 +26,13 @@ item_related_router.register(
     basename="invitations",
 )
 
+sdk_relay_router = DefaultRouter()
+sdk_relay_router.register(
+    "sdk-relay/events",
+    viewsets.SDKRelayEventViewset,
+    basename="sdk_relay_events",
+)
+
 
 urlpatterns = [
     path(
@@ -38,14 +45,7 @@ urlpatterns = [
                     r"^items/(?P<resource_id>[0-9a-z-]*)/",
                     include(item_related_router.urls),
                 ),
-                re_path(
-                    r"^sdk-relay/events/(?P<token>[0-9a-zA-Z-]*)/",
-                    viewsets.SDKRelayEventView.as_view(),
-                ),
-                re_path(
-                    r"^sdk-relay/events/",
-                    viewsets.SDKRelayEventCreateView.as_view(),
-                ),
+                *sdk_relay_router.urls
             ]
         ),
     ),